allow to remove annotations from logfile
in Jenkins file runner we observed a race condition between the build
writing the log and JFR reading the log to print to stdout.
Every now and then it seems the annotations are not fully written,
leading to exceptions and the main thread of JFR dies but as other
threads are not terminated JFR effectively hangs.