Apply Sonar finding "Change "try" to "try-with-resources"
As per Java doc it is more efficient to use try-with-resources than
having finally block and close the objects

```
However, this example might have a resource leak. A program has to do
more than rely on the garbage collector (GC) to reclaim a resource's
memory when it's finished with it. The program must also release the
resoure back to the operating system, typically by calling the
resource's close method. However, if a program fails to do this before
the GC reclaims the resource, then the information needed to release the
resource is lost. The resource, which is still considered by the
operaing system to be in use, has leaked.
```

See
https://docs.oracle.com/javase/tutorial/essential/exceptions/tryResourceClose.html