Move config params for direct writer
When the ElasticSearchWriteAccess extension point was introduced, the
configuration parameters for ElasticSearchWriteAccessDirect have been
left in the common configuration (ElasticSearchConfiguration) to stay
compatible.

This is not intuitive for users and makes the code more complicated.

Move all configuration parameters specific to
ElasticSearchWriteAccessDirect to this class.

Also rename configuration parameters to better reflect their meanings.

This is an incompatible change! Users must adapt their plugin
configuration.