Run fluentd.emitData in a separate thread
As the main threads are affected by the Durable tasks and interrupted
exceptions can occur due to long sleeps it was decided to run the
fluentd.emitData in a separate thread.

Initialized new newCachedThreadPool which can re-use the thread in case
there exists an idle thread in the threadpool, otherwise creates a new
thread and inject it to the threadpool. As thread creation is costly
re-using threads helps the performance.