Run emitting steps of the fluency library in a separate thread (#108)
Threads might get interrupted when stopped by the
'workflow-durable-task-step-plugin'. Such interruptions does not allow
fluency to properly flush the queued events in the buffer.
java.io.IOException: Failed to move retention buffer due to interruption

See interruptions in the 'workflow-durable-task-step-plugin' plugin.
https://github.com/jenkinsci/workflow-durable-task-step-plugin/blob/master/src/main/java/org/jenkinsci/plugins/workflow/steps/durable_task/DurableTaskStep.java#L392
https://github.com/jenkinsci/workflow-support-plugin/blob/master/src/main/java/org/jenkinsci/plugins/workflow/support/concurrent/Timeout.java#L86

As the main threads are affected by the Durable tasks and interrupted
exceptions can occur due to long sleeps it was decided to run the
fluentd.emitData in a separate thread.

Initialized new newCachedThreadPool which can re-use the thread in case
there exists an idle thread in the threadpool, otherwise creates a new
thread and inject it to the threadpool. As thread creation is costly
re-using threads helps the performance.


Co-authored-by: Andreas Brehmer <andreas.brehmer@sap.com>